Fix warning from upd_0100.php during update #6894

Closes #6894

See merge request ispconfig/ispconfig3!2036

+++ install/patches/upd_0100.php
@@ -8,9 +8,9 @@ class upd_0100 extends installer_patch_update {
 		global $inst;
 
         // Remove old server plugins, unless they are currently enabled
-        if(!is_link('/usr/local/ispconfig/server/plugins-enabled/nginx_reverseproxy_plugin.inc.php'))
+        if(file_exists('/usr/local/ispconfig/server/plugins-available/nginx_reverseproxy_plugin.inc.php') && !is_link('/usr/local/ispconfig/server/plugins-enabled/nginx_reverseproxy_plugin.inc.php'))
             unlink('/usr/local/ispconfig/server/plugins-available/nginx_reverseproxy_plugin.inc.php');
-        if(!is_link('/usr/local/ispconfig/server/plugins-enabled/bind_dlz_plugin.inc.php'))
+        if(file_exists('/usr/local/ispconfig/server/plugins-available/bind_dlz_plugin.inc.php') && !is_link('/usr/local/ispconfig/server/plugins-enabled/bind_dlz_plugin.inc.php'))
             unlink('/usr/local/ispconfig/server/plugins-available/bind_dlz_plugin.inc.php');
 	}
